lol, "fixed" the pause menu problem
Turns out the solution was the entry task today lol, and I did it in my last push

Instead, to make it seem like I actually did something, I fixed a bug that sent the player to a scene that wasn't in the build settings (probably because this is copied and pasted code). The Main Menu button now sends the player to the start screen.

As a side note, if anyone wants to work on the UI buttons, they still have default texturing