I guess I haven't saved in a while, so save push I guess
Still working on the AI. I found a new tutorial that suits the situation better, so I updated it a bit. Also, for whatever reason, the AI tabs in the Windows were removed and placed in a package, which I admit is where most of my time went trying to figure that out lol. So, a new package has been downloaded.