I think I figured out the NavMesh problem
Are you guys ready for the solution to why the NavMesh wasn't working, keeping me up all night and putting me on 4 hours of sleep today? It was... I was an idiot. It was working, it's just really slow building it and the enemy moves really slowly.

But seriously, I am on 4 hours of sleep today. I need a break.