Implementing a very bad collectible pickup system
NOW WITH RAYCASTS!!!
(In case it wasn't obvious, I'm glad we're actually visiting the coding sections. I've heard a little bit about the raycasts, and they are pretty powerful)
Anyways, the system is going to use Raycasts instead of collisions to get the collectibles. I might be leaving soon, so this is also like an end of the day push, and I just wanted to say that if it gives you errors, feel free to go in and comment all of it out to stop them lol. It's very bad now, mostly because I haven't tested it at all.