Oopsie... we're not leaving early
That's right, you have to deal with me and Mitchell for another half an hour heh heh heh...
I'm kidding. I am, of course, still working on the collectibles, where I corrected the system errors a bit. Raycasts are a little wierd, but I think I can get used to them. I think I'm almost ready to import a very basic sample collectible into one of the scenes.
Now, I think I have a new problem, specifically for the designers. The collectibles can look like whatever, I don't care, I'm assuming that because this is a research facility, there is more, less important research around (basically to dress the scene), and they need to be distinct from each other, so the player knows what to go for. This is a task for you guys; I'll repeat it on the trello board, because I know I mostly write descriptions on GitHub for myself.