Level 2, and all the AI, should be done
Again, it's done from a technical standpoint, not a level design one. Someone else can feel free to tell me to redesign. I believe that means that I'm done, so I'm going to spend the rest of my time bug testing and probably fixing the many errors that I made.