Fixed the NavMesh bug
Mitchell found that the enemy would wander in circles around the player, because the angle speed (turning speed) was too low and the enemy would basically start orbiting instead of tracking.
It's fixed now. The value in the agent is 360, which is complete snap turns. If someone wants to lower it, feel free to for an easier or more natural experience.